% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_os}
\alias{get_os}
\title{Get Operating System}
\usage{
get_os()
}
\value{
Return any of the following:
\describe{
\item{"linux"}{GNU/Linux}
\item{"osx"}{Macintosh OSx}
\item{"windows"}{Windows}
\item{mystery machine}{Operating systems other than the three above}
}
}
\description{
Utility to detect the operating system of the current \code{R} session.
Based on \url{http://conjugateprior.org/2015/06/identifying-the-os-from-r/}.
}
\examples{
get_os()
}
